pyproject.toml
[project]
name = "astraia"
version = "0.1.0"
description = "LLM-driven optimization framework (MVP skeleton)."
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "pyyaml>=6.0",
  "optuna>=3.6",
  "pydantic>=2.5",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "httpx>=0.27",
  "matplotlib>=3.8",
  "pandas>=2.2",
  "qiskit>=1.0",
  "hypothesis>=6.100",
  "qiskit-qasm3-import>=0.5",
]
authors = [
  { name = "Astraia Dev Team" }
]
license = { text = "MIT" }

[project.optional-dependencies]
openai = ["openai>=1.0"]
gemini = ["google-generativeai>=0.5"]
llm = ["openai>=1.0", "google-generativeai>=0.5"]
quantum = ["qiskit>=1.0", "qiskit-qasm3-import>=0.5"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
astraia = "astraia.cli:main"
